+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Deleting an organ in the Organ tab of the Settings dialog causes also deleting it's cache and all it's presets https://github.com/GrandOrgue/grandorgue/issues/1049
 - Reorganised the Organs tab of the Settings dialog https://github.com/GrandOrgue/grandorgue/issues/1049
 - Added capability of restoring organ settings values for all subobjects https://github.com/GrandOrgue/grandorgue/issues/1129
 # 3.6.7 (2022-05-15)